// app/src/pages/project/__generated__/TraceErrorsTimeSeriesQuery.graphql.ts
export type TimeRange = {
  end?: string | null;
  start?: string | null;
};
export type TimeBinConfig = {
  scale?: TimeBinScale;
  utcOffsetMinutes?: number;
};
export type TraceErrorsTimeSeriesQuery$variables = {
  projectId: string;
  timeBinConfig: TimeBinConfig;
  timeRange: TimeRange;
};
export type TraceErrorsTimeSeriesQuery$data = {
  readonly project: {
    readonly traceCountByStatusTimeSeries?: {
      readonly data: ReadonlyArray<{
        readonly errorCount: number;
        readonly timestamp: string;
      }>;
    };
  };
};
export type TraceErrorsTimeSeriesQuery = {
  response: TraceErrorsTimeSeriesQuery$data;
  variables: TraceErrorsTimeSeriesQuery$variables;
};